EduMUSE-tts/multi_agent_pipeline/requirements.txt
crewai
openai
whisper
pdfminer.six
beautifulsoup4
requests
numpy
gTTS
torch
torchaudio
tortoise-tts
elevenlabs
pydub
python-dotenv
